refactor(prism/review): split review.go (2071 LoC) into 9 per-concern files
Pure structural split with no behaviour changes. The 2071-line monolith is
replaced by nine files, each owning a single concern:

- review.go    — package doc, constants, core types (PRContext, Opts, etc.)
- context.go   — FetchPRContextWithOpts and GitHub/git integration helpers
- agents.go    — Agent catalogue, AgentsByName, CheckAgentAvailability, ResolveAgentConfigContent
- run.go       — Run(), RunAsync(), buildAsyncAck(), LookupParentSession()
- poll.go      — pollAgents(), buildResults(), BuildResults(), isTerminalState()
- results.go   — FormatResults(), AssessPassed(), FindingsFilePath(), failureReason()
- prompt.go    — buildReviewPrompt(), sortStrings(), deriveRepo(), formatDuration()
- lifecycle.go — NextRoundNumber(), KillSession* variants, CleanupReviewSessionsForParent()
- export_test.go — *ForTest shims relocated from review.go (_test.go convention)

readiness.go is unchanged. go build ./... and go test ./... pass.

Closes #1153
Refs #1090
